extract texera_blocking_unsupervised_trainer_operator.py
