+++ core/amber/src/main/resources/python_udf/operators/texera_udf_operator_base.py
@@ -92,35 +92,3 @@ class TexeraUDFOperator(ABC):
         Executes when the input is exhausted, useful for some blocking execution like training.
         """
         pass
-
-
-class TexeraBlockingUnsupervisedTrainerOperator(TexeraUDFOperator):
-    logger = logging.getLogger("PythonUDF.TexeraBlockingUnsupervisedTrainerOperator")
-
-    @exception(logger)
-    def __init__(self):
-        super().__init__()
-        self._data = []
-        self._train_args = dict()
-
-    @exception(logger)
-    def accept(self, row: pandas.Series, nth_child: int = 0) -> None:
-        self._data.append(row[0])
-
-    @exception(logger)
-    def close(self) -> None:
-        pass
-
-    @staticmethod
-    @exception(logger)
-    def train(data, *args, **kwargs):
-        raise NotImplementedError
-
-    @exception(logger)
-    def report(self, model) -> None:
-        pass
-
-    @exception(logger)
-    def input_exhausted(self, *args):
-        model = self.train(self._data, **self._train_args)
-        self.report(model)
